+++ src/components/MediatedRequestsActivities/components/index.js
@@ -1,0 +1,2 @@
+export { default as MediatedRequestsFilters } from './MediatedRequestsFilters';
+export { default as MediatedRequestsList } from './MediatedRequestsList';