+++ src/constants/index.js
@@ -1,3 +1,3 @@
 export * from './base';
 export * from './api';
-export * from './filters';
+export * from './mediatedRequestsActivities';