+++ test/jest/__mock__/index.js
@@ -1,3 +1,4 @@
+import './currencyData.mock';
 import './stripesConfig.mock';
 import './stripesCore.mock';
 import './stripes.mock';