+++ test/jest/__mock__/stripesSmartComponents.mock.js
@@ -12,16 +12,25 @@ jest.mock('@folio/stripes/smart-components', () => ({
       </button>
     </div>
   )),
-  CheckboxFilter: jest.fn((props) => (<div {...props} />)),
+  CheckboxFilter: jest.fn(({
+    'data-testid': testId,
+    onChange,
+    ...rest
+  }) => (
+    <div
+      {...rest}
+      data-testid={testId}
+    >
+      <button
+        type="button"
+        onClick={onChange}
+        data-testid={`${testId}Button`}
+      >Change
+      </button>
+    </div>
+  )),
+  SearchAndSortNoResultsMessage: jest.fn((props) => (<div {...props} />)),
   ViewMetaData: jest.fn(({ metadata, ...rest }) => (
     <div {...rest}>{metadata.createdDate}</div>
   )),
-  SearchAndSortQuery: ({
-    children,
-    'data-testid': testId,
-  }) => (
-    typeof children === 'function'
-      ? <div data-testid={testId}>{ children() }</div>
-      : <div data-testid={testId}>{ children }</div>
-  ),
 }));