+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Add Filters for Mediated requests actions page. Refs UIREQMED-4.
 * Update permission for Mediated requests. Refs UIREQMED-29.
 * Implement `Create mediated request` form with basic functionality. Refs UIREQMED-27.
+* Add Mediated request status and Mediated request level filters for Mediated requests actions page. Refs UIREQMED-25.
+* Add Search for Mediated requests actions page. Refs UIREQMED-26.
 
 ## 1.0.0
 * New app created with stripes-cli. Updated module after created with stripes-cli. Refs UIREQMED-1.